refactor record hook
