+++ src/components/ProductCardHeader/index.tsx
@@ -5,6 +5,7 @@ import BlueRing from 'components/Icon/BlueRing'
 import { CURRENCIES } from 'constants/currencies'
 import useBreakpoint from 'hooks/useBreakpoint'
 import { usePrice } from 'hooks/usePriceSet'
+import { toLocalNumberString } from 'utils/toLocalNumberString'
 
 interface Props {
   logoCurSymbol: string
@@ -65,12 +66,7 @@ export default function ProductCardHeader(props: Props) {
           gap={isDownMd ? 10 : 0}
         >
           <Typography color="primary" fontSize={24} fontWeight={700} gap={8} display="flex" alignItems="center">
-            <span style={{ width: 120 }}>
-              {' '}
-              {curPrice
-                ? (+curPrice).toLocaleString('en-US', { minimumFractionDigits: 2, maximumFractionDigits: 2 })
-                : '-'}
-            </span>
+            <span style={{ width: 120 }}> {curPrice ? toLocalNumberString(curPrice) : '-'}</span>
             <BlueRing />
           </Typography>
           <Typography fontSize={16} sx={{ color: theme => theme.palette.text.secondary }}>
