+++ src/hooks/useSuccessImage.tsx
@@ -3,7 +3,7 @@ import { useCallback, useMemo, useState } from 'react'
 import dayjs from 'dayjs'
 import CryptoJs from 'crypto-js'
 import { Axios } from 'utils/axios'
-import { OrderRecord } from 'utils/fetch/product'
+import { OrderRecord } from 'utils/fetch/record'
 import useModal from './useModal'
 import usePollingWithMaxRetries from './usePollingWithMaxRetries'
 import { CURRENCY_ADDRESS_MAP } from 'constants/currencies'
