+++ src/pages/Account/Dashboard.tsx
@@ -26,6 +26,7 @@ import { usePrice } from 'hooks/usePriceSet'
 import { useAccountBalances } from 'hooks/useAccountBalance'
 import { toChecksumAddress } from 'web3-utils'
 import { CURRENCIES } from 'constants/currencies'
+import { toLocalNumberString } from 'utils/toLocalNumberString'
 
 enum BalanceTableHeaderIndex {
   token,
@@ -282,19 +283,13 @@ export default function Dashboard() {
 
                 {isDownMd ? (
                   <InvestmentValueCard
-                    value={(+totalInvest).toLocaleString('en-US', {
-                      minimumFractionDigits: 6,
-                      maximumFractionDigits: 6
-                    })}
+                    value={toLocalNumberString(totalInvest, 6)}
                     unit="$"
                     // dayChange="+ 8.91% / $350.28 "
                   />
                 ) : (
                   <NumericalCard
-                    value={(+totalInvest).toLocaleString('en-US', {
-                      minimumFractionDigits: 6,
-                      maximumFractionDigits: 6
-                    })}
+                    value={toLocalNumberString(totalInvest, 6)}
                     border
                     title="Portfolio Value"
                     unit="$"
