+++ src/pages/ChainOption/index.tsx
@@ -17,6 +17,7 @@ import { useParams } from 'react-router-dom'
 import ProductBanner from 'components/ProductBanner'
 import BlueRing from 'components/Icon/BlueRing'
 import { SUPPORTED_CURRENCY_SYMBOL } from 'constants/currencies'
+import { toLocalNumberString } from 'utils/toLocalNumberString'
 
 enum TYPE {
   Saddle = 'Saddle',
@@ -134,9 +135,7 @@ export default function ChainOption() {
       <ProductBanner
         title="Chain-type Option"
         checkpoints={['Easy to access，enjoy high returns']}
-        val1={
-          BTCPrice ? (+BTCPrice).toLocaleString('en-US', { minimumFractionDigits: 2, maximumFractionDigits: 2 }) : '-'
-        }
+        val1={BTCPrice ? toLocalNumberString(BTCPrice) : '-'}
         subVal1={'BTC latest spot price'}
         unit1={
           <Box display={'flex'} alignItems="center" gap={'5px'} component="span">
@@ -144,9 +143,7 @@ export default function ChainOption() {
             <BlueRing />
           </Box>
         }
-        val2={
-          ETHPrice ? (+ETHPrice).toLocaleString('en-US', { minimumFractionDigits: 2, maximumFractionDigits: 2 }) : '-'
-        }
+        val2={ETHPrice ? toLocalNumberString(ETHPrice) : '-'}
         subVal2={'ETH latest spot price'}
         unit2={
           <Box display={'flex'} alignItems="center" gap={'5px'} component="span">
