+++ src/pages/RecurringVault/VaultProductCard.tsx
@@ -8,13 +8,14 @@ import { OutlinedCard } from 'components/Card/Card'
 import { RecurProduct } from 'utils/fetch/recur'
 import Spinner from 'components/Spinner'
 import dayjs from 'dayjs'
+import { useRecurPnl } from 'hooks/useRecurData'
+import { toLocalNumberString } from 'utils/toLocalNumberString'
 
 export default function VaultProductCard({
   logoCurSymbol,
   priceCurSymbol,
   title,
   description,
-  deposit,
   onClick,
   color,
   product
@@ -23,11 +24,11 @@ export default function VaultProductCard({
   priceCurSymbol: string
   title: string
   description: string
-  deposit: string
   onClick: () => void
   color: string
   product: RecurProduct | undefined
 }) {
+  const { totalReInvest } = useRecurPnl(product?.investCurrency)
   return (
     <Box
       display="grid"
@@ -53,7 +54,11 @@ export default function VaultProductCard({
       {product ? (
         <Box display={{ xs: 'grid' }} gap={24} gridTemplateColumns={{ xs: '100%', md: '35% auto' }}>
           <NumericalCard value={product?.apy ?? '-'} subValue="Current APY" border valueColor={color} />
-          <NumericalCard value={deposit} subValue="Your existing position" border>
+          <NumericalCard
+            value={toLocalNumberString(totalReInvest) + (product?.investCurrency ?? '-')}
+            subValue="Your existing position"
+            border
+          >
             <Button
               style={{ borderRadius: 40 }}
               width={'132px'}
@@ -64,7 +69,11 @@ export default function VaultProductCard({
               Add
             </Button>
           </NumericalCard>
-          <NumericalCard value={product.strikePrice + ' USDT'} subValue="Current Strike Price" border />
+          <NumericalCard
+            value={toLocalNumberString(product.strikePrice) + ' USDT'}
+            subValue="Current Strike Price"
+            border
+          />
           <NumericalCard
             value={<Timer timer={product?.expiredAt ?? 0} />}
             subValue="Countdown until delivery date"
