+++ src/utils/fetch/product.ts
@@ -5,22 +5,6 @@ const TYPE = {
   put: 'PUT'
 }
 
-export enum InvesStatusType {
-  SUCCESS = 'success',
-  PENDING = 'pending',
-  ERROR = 'error'
-}
-
-export const InvesStatus = {
-  [1]: InvesStatusType.PENDING,
-  [2]: InvesStatusType.SUCCESS,
-  [3]: InvesStatusType.SUCCESS,
-  [4]: InvesStatusType.SUCCESS,
-  [5]: InvesStatusType.ERROR,
-  [6]: InvesStatusType.PENDING,
-  [7]: InvesStatusType.ERROR
-}
-
 export interface createOrder {
   address: string
   amount: number
@@ -73,35 +57,6 @@ export type SingleCurProductList = { call: Product[]; put: Product[] }
 
 export type ProductList = {
   [key in Partial<keyof typeof SYMBOL_MAP>]: SingleCurProductList
-}
-
-export interface OrderRecord {
-  address: string
-  amount: number
-  annualRor: string
-  confirmOrderHash: string
-  createdAt: number
-  currency: string
-  deliveryPrice: string
-  earn: string
-  expiredAt: number
-  hash: string
-  indexPrice: string
-  investStatus: number
-  investCurrency: string
-  isLiquidated: string
-  multiplier: string
-  orderId: number
-  price: string
-  productId: number
-  returnedAmount: string
-  returnedCurrency: string
-  signCount: string
-  status: string
-  strikeCurrency: string
-  strikePrice: string
-  ts: number
-  type: string
 }
 
 export const productFormatter = (raw: ProductRaw): Product => {
